server/requirements.txt
anyio==3.3.4
asgiref==3.4.1
black==21.10b0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
motor==2.3.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.1
python-decouple==3.5
regex==2021.11.10
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
